Refactoring of URL's and email confirmation.
What used to be email verification is now called email confirmation.
Also refactored some URL's that were user specific but were not opened
under ``accounts/<username/..``. Now activating and confirming an email
address is account specific. This makes it even more secure because an
unqiue SHA1 is checked in combination with the username.