Update utils.py
Previously the herein referenced AUTH_PROFILE_MODULE could not point to a model for a subpackaged app because this get_profile_model function splits on all periods. As we know get_model takes app_label and model as arguments so for a subpackaged app this would break. This correctly interprets the app_label and model by splitting on the last period in the string.