Tox: Install Django 1.4.x from source
Recent versions of pip (7.x+) default to creating and using .whl packages.
In the case of Django 1.4 this installs the templates to the root of a
virtualenv, rather than '.../site-packages/django/...'.

So when userna's tests are run inside Tox any that use templates fail with

    TemplateDoesNotExist: registration/password_reset_subject.txt

installing Django 1.4 from source fixes this.